Improvement of centered text rendering
drawCenterText now works with line wraps. It respects maxium avaiable width but not maximum height. The x and y parameters denote the center of the text, it grows in both directions.
Further improvement should be one method for all text rendering with a bounding box and text alignment parameter in that box.