fix issue: Plugging a Countour Shuttle into a different USB port treats it as a different device
Completes issue #3486

This PR assigns an ID in the order the Contour Shuttle is encountered. In the majority of cases, the user will have only one Contour Shuttle (of a particular type) and this will fully resolve any issues. If the edge-case in which a user has two devices of the same type, and they are assigned in a different order than the user wants, the user can simply unplug both devices and reattach them in the desired order. (Or use surface groups)