Improve new `@use` workflow
This adjusts the code from PR #3523 to improve the workflow and clarity, and to add back a minor feature that had been removed as a consequence of that PR.

- Change the `@use...with` in *_variables.scss* to `@forward...with`

-- This allows all variables to be loaded with: `@use 'variables'` in the other subfiles, making the origin of the variable-values clearer. (As opposed to needing to load `@coreui` even though the overrides were defined in *_variables.scss*. )

-- It also allows using `!default` as it was used in the original _variables.scss.

- Change remaining `@use '@coreui'` rules to @use 'variables'`

note: I left an `@use '@coreui'` in *App.scss* for clarity only. It doesn't actually do anything -- good or bad.
Likewise, adding `@use '@coreui'` to allow identifying mixins that originate in coreui