fix: Edit Button Load PNG not loading updated files
When the contents of a PNG file is updated, replacing the PNG image from the file of the same name doesn't always work. Specifically,
1. Load a PNG file.
2. Stay on the Buttons page
3. Modify the PNG file.
4. Try to replace the PNG file for the button.

Moving to a different buttons page doesn't help, though moving to a different URL or restarting will solve the problem.

This commit fixes the issue, so one can do rapid "protoyping" of the PNG image without leaving the "Buttons" page.

While in theory, one should test for when to erase the existing value, doing it all the time doesn't appear to harm anything even if the user cancels the file dialog.