Fix:Preset previewStyle is overridden by feedback
Issue #3669

This is what I meant by my "proposal" in the comments and I think it's also what you are suggesting, but if not then it's a point of departure for discussion.

While one could try to be more clever about what one is overriding, it will generally fail with "advanced" feedbacks, since the style information is missing then.