Allow the `+` and `+=` expression operators to act on strings
So, for example: 'Fader: ' + toFixed($(module:faderValue),1) returns 'Fader: 1.0' rather than N/A

Note: This implementation will ignore trimmable spaces, so the user cannot force it to be a string-op by starting with space or empty string, i.e.

   `' ' + 7 + 5` equals 12 (number) and not '75' (string)

It could be fixed with a bit of effort, but is probably rare enough to ignore?