Fix: Export is saving twice
The form has two submit callbacks: removed one of them.

Note that the dev environment seems a bit more fragile -- once there's a timeout, the buttons stop saving. This may be something on the server side or in the communication method, since the UI side appears to be working correctly.