Fix handling of variables in button Template Literals
This PR simplifies handling of Companion variables in button text-expressions so the user can write `Prefix_$(custom:var)` rather than `Prefix_${$(custom:var)}

Details: Currently button text expressions have a curious behavior:

If you put a companion variable in a Template Literal, the variable is not evaluated, so to evaluate it you'd have to write: `${$(my:var)}`, which is rather awkward.

OTOH, if you set a variable, say $(custom:myvar), to that same template literal, i.e. `$(my:var)`, and then use the new variable as the button text value (whether expression or text), the button now shows the _value_ of $(my:var) rather than the string "$(my:var)" -- even though the string is shown as the value in the Variables tab.

This PR fixes the immediate problem of simplifying template literal syntax in button expressions and also suggests an alternate option, which is to resolve backtick expressions at parse-eval time, just like it already done for rest of the expression. (i.e. if the expression is `unixNow()`, that function is evaluated only immediately after parsing.)