Feature: Two-line entries for variable-name dropdowns
This PR makes the variable-name dropdown menus behave in actions and feedbacks similarly to the way they work in the new expression editor:

Each entry in the dropdown is two lines:
~~~
$(module:varname)
Variable description text
~~~

NB: The main effect of the CSS change in this PR is to allow `\n` to break lines in menu-items. In addition, it removes some "excessive" spacing between menu items.

Wrapping behavior is not changed: Prior to this PR long values were wrapped in menu items but not in the current value (i.e. when dropdown is not "open). This PR does not change the behavior, but it would be very easy to wrap text even in the current value -- in fact I had to do extra work to prevent it!

Conversely, if preferred: with a little effort the CSS changes in this PR can be isolated to affect only variable dropdowns.