chore: require double-quotes in YAML
Consider this PR as a suggestion...

Currently prettier config requires single-quotes in YAML files, but double-quotes are probably preferred since it allows escape sequences.  (See for example [Github's own examples](https://docs.github.com/en/actions/how-tos/write-workflows/choose-what-workflows-do/find-and-customize-actions#using-inputs-and-outputs-with-an-action).)

This PR removes the need for excluding _.yarnrc.yml_ but does require "reformatting" the github workflow files (done in this PR).

FWIW, prettier preserves single quotes if it would change the meaning. For example, if you have a value such as `'\n means EOL'`, prettier will leave the single-quotes.