feat: Improve visibility of the user guide
A recurring question on FB is some variation on "where is the user guide?" (Even I needed to be shown the sidebar link.) This PR puts the link top-right in the unused space of the header bar. It does not take up any new space, and frees up the space on the sidebar.

I also changed the icon from "i" to "?" for improved visibility, and note that, in response to query: "best icon for help", Google says "Question Mark (?): The classic symbol for general help, FAQs, or tutorials."

This PR also fixes a "bug" in which the text seems to disappear after the user clicks on a link in the header (text turns dark on focus).