Fix: "Unset" surface variable "breaks" equipment
v4.2 eliminated the internal jog/shuttle/t-bar variables in favor of user-specified custom variables. While this is a clear improvement in behavior, the default of "none" (no variable) causes "null" values to be sent to my mixer, which caused it to fail internally (needed reboot to restore full functionality).

This PR restores the now-deprecated internal variables in the case that 'none' is selected. (A very minor side-effect: a variable will always be set, even if 'none' is selected. Since this corresponds to the behavior in v4.1, it shouldn't be an issue.)

Note that this PR (a) is a hotfix for v4.2 and (b) restores v4.2 to be non-breaking (at least in this regard) as might be expected in SemVer....

I will leave it to Julian to decide whether this should be implemented in v4.3 (i.e. `main`) as well, where surfaces are in independent modules. (And whether he wants me to do it!)

Note 2: I tested this on a Contour Shuttle and assume it will therefore work with the XKeys and Loupedecks as well.

Note 3: BlackmagicController is not reverted here because it was already using variables prior to v4.2