Fix invalid values being sent to a surface
This restores the internal t-bar/jog/shuttle variables and marks them as deprecated. The variables are never updated but exist to prevent null values being sent to a surface.  See PR #3869 for details.

NOTE: I am posting this as a convenience, not a nudge! If you disagree with my proposal in [my comment on 3869](https://github.com/bitfocus/companion/pull/3869#issuecomment-3697186142) just ignore or close this PR.